refactor(core): replace long with Int type for array and string operations
- Changed Array::slice method parameters from long to Int type
- Updated Extension::registerConstant parameter type from long to Int
- Modified prepare_slice function signature to use Int instead of long
- Replaced String::substr method parameters from long to Int type
- Updated printf format specifiers to use ZEND_LONG_FMT macro
- Changed Variant class printf output to use ZEND_LONG_FMT format

+++ include/phpx_ext.h
@@ -264,7 +264,7 @@ class Extension {
     bool registerInterface(Interface *i) const;
     bool registerFunctions(const zend_function_entry *functions);
     bool registerResource(const char *name, resource_dtor dtor) const;
-    void registerConstant(const char *name, long v) const;
+    void registerConstant(const char *name, Int v) const;
     void registerConstant(const char *name, int v) const;
     void registerConstant(const char *name, bool v) const;
     void registerConstant(const char *name, double v) const;
